requirements.txt
# Core deep learning
torch>=1.12.0
torchvision>=0.13.0

# Numerical and image I/O
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0
scikit-image>=0.19.0

# Random Forest & metrics
scikit-learn>=1.0.0
joblib>=1.2.0

# Config & progress
PyYAML>=6.0
tqdm>=4.64.0

# Visualization & reporting
matplotlib>=3.5.0
pandas>=1.3.0
reportlab